decoupling the samplers from zooniverse
