+++ ChildProject/cmdline.py
@@ -242,6 +242,7 @@ def compute_durations(args):
     recordings.to_csv(os.path.join(project.path, 'metadata/recordings.csv'), index = False)
 
 def main():
+    register_pipeline('sampler', SamplerPipeline)
     register_pipeline('zooniverse', ZooniversePipeline)
     register_pipeline('convert', ConversionPipeline)
