+++ ChildProject/pipelines/__init__.py
@@ -1,2 +1,3 @@
+from .sampler import SamplerPipeline
 from .zooniverse import ZooniversePipeline
 from .conversion import ConversionPipeline
